chatbot-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sentence-transformers==2.2.2
numpy==1.24.3
firebase-admin==6.2.0
python-dotenv==1.0.0
huggingface-hub==0.16.4
torch==2.0.1
transformers==4.30.2
pillow==10.0.0
scikit-learn==1.3.0
pandas==2.0.3
opencv-python==4.8.0.76
python-multipart==0.0.9 